Remove GenericArticleTask
